requirements.txt
opencv-python~=4.8.1.78
numpy~=1.26.0
torch~=2.3.0
albumentations~=1.4.21
matplotlib~=3.8.0
Pillow~=10.0.1